do auth server again, good practice
