+++ lib/commands/auth/login.js
@@ -36,7 +36,7 @@ function login (opts) {
     client.login({
       email: user.email,
       password: user.password
-    }, function (err) {
+    }, function (err, resp, body) {
       if (err) ui.exitErr(err)
       console.log('Logged in successfully.')
       process.exit(0)
