+++ package.json
@@ -30,11 +30,12 @@
     "rimraf": "^2.5.4",
     "status-logger": "^3.0.0",
     "subcommand": "^2.0.4",
-    "township-client": "^1.2.1",
+    "township-client": "^1.3.1",
     "xtend": "^4.0.1"
   },
   "devDependencies": {
     "appa": "^5.0.0",
+    "dat-land": "datproject/datfolder",
     "hypercore": "^4.15.1",
     "hyperdiscovery": "^1.0.1",
     "memdb": "^1.3.1",
