+++ tests/helpers/auth-server.js
@@ -1,33 +1,41 @@
-var memdb = require('memdb')
-var http = require('http')
-var createAppa = require('appa')
-var township = require('township')
+var path = require('path')
+var Server = require('dat.land/server')
+var initDb = require('dat.land/server/database/init')
+var rimraf = require('rimraf')
 
 module.exports = createServer
 
 function createServer (port, cb) {
-  var app = createAppa({log: {level: 'silent'}})
-  var db = memdb()
-  var ship = township({secret: 'test'}, db)
+  var config = {
+    township: {
+      secret: 'very secret code',
+      db: path.join(__dirname, '..', 'test-township.db')
+    },
+    db: {
+      dialect: 'sqlite3',
+      connection: { filename: path.join(__dirname, '..', 'test-sqlite.db') },
+      useNullAsDefault: true
+    },
+    whitelist: false,
+    port: port || 8888
+  }
 
-  app.on('/register', function (req, res, ctx) {
-    // appa provides `ctx` for us in the way we want out of the box
-    ship.register(req, res, ctx, function (err, code, data) {
-      if (err) return app.error(res, code, err.message)
-      app.send(res, code, data)
+  initDb(config.db, function (err, db) {
+    if (err) return cb(err)
+
+    const server = Server(config, db)
+    server.listen(config.port, function () {
+      console.log('listening', config.port)
     })
-  })
 
-  app.on('/login', function (req, res, ctx) {
-    // appa provides `ctx` for us in the way we want out of the box
-    ship.login(req, res, ctx, function (err, code, data) {
-      if (err) return app.error(res, code, err.message)
-      app.send(res, code, data)
-    })
-  })
+    cb(null, server, close)
 
-  var server = http.createServer(app)
-  server.listen(port, function () {
-    cb(server)
+    function close (cb) {
+      server.close(function () {
+        rimraf.sync(config.township.db)
+        rimraf.sync(config.db.connection.filename)
+        process.exit()
+      })
+    }
   })
 }
